remove redundant page components and put them directly in to page
