--- frontend/src/components/Home/HomeDesktop.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './home.css';
-import { Button, Typography } from '@mui/material';
-
-export default function HomeDesktop() {
-  return (
-    <div>
-      <Typography variant="h1">HOME PAGE</Typography>
-      <Button>Hello World</Button>
-    </div>
-  );
-}
