--- frontend/src/components/Home/HomeMobile.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import './home.css';
-import { Button, Typography } from '@mui/material';
-
-export default function HomeDesktop() {
-  return (
-    <div id="wrapper1">
-      <div>
-        <Typography variant="h1">HOME PAGE</Typography>
-        <Button>Hello World</Button>
-      </div>
-    </div>
-  );
-}
