+++ frontend/src/pages/landing.tsx
@@ -3,8 +3,8 @@ import { Typography } from '@mui/material';
 
 export default function LandingPage() {
   return (
-    <div>
+    <>
       <Typography variant="h1">LANDING PAGE</Typography>
-    </div>
+    </>
   );
 }
