--- frontend/src/pages/login.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import LoginComponent from '../components/Login/index';
-
-class LoginPage extends React.Component {
-  render() {
-    return <LoginComponent />;
-  }
-}
-
-export default LoginPage;
