--- frontend/src/pages/projects.ts
@@ -1,6 +1,0 @@
-import React from 'react';
-import DisplayProjects from '../components/DisplayProjects';
-
-export default function ProjectsPage() {
-  return React.createElement(DisplayProjects);
-}
